Fix GHA workflow by updating gemfile.lock
The GHA workflow was failing at the `bundler install` stage due to
changes in the gemfile/gemspec not being represented in the Gemfile.lock
file. To fix this issue, I had bundler regenerate the lockfile.

To fix install problems that I ran into while trying to update the
lockfile, I updated the FFI version and removed the version restriction
on bundler.

DSET-2194