Fix test_bad_ssl_certificates test by adding a bad ca cert
Newer ruby versions raise an exception when a user tries to load a path
to a cert (or path) that doesn't exist or is invalid. In order to get
the `test_bad_ssl_certificates` test working, I created a dummy cert
with a different hostname so we would get the failure the test is
looking for.

DSET-2194